requirements.txt
django==3.0.4
django-allauth==0.41.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-filters==1.0.0.dev0
PyMySQL==0.9.3
django-gcloud-storage==0.4.0
django-storages==1.9.1
django-mysql==3.8.1
elasticsearch==7.8.1
django-debug-toolbar==2.2
django-filter==2.3.0
Pillow==8.0.1
channels==2.4.0
channels-redis==2.4.2
drf-yasg==1.20.0
daphne==2.5.0
gunicorn==20.1.0
whitenoise==5.2.0
django-gmailapi-backend==0.3.0
django-invitations==1.9.3
https://github.com/darklow/django-suit/tarball/v2